bch(m5): merkle-root accept gate before full_block emit; share merkle.hpp
Validate every received/reconstructed block against its header merkle
commitment before publishing it on full_block. All three delivery paths
(direct block msg, compact-block-complete, compact+getblocktxn round-trip)
now funnel through emit_full_block(), which recomputes the SHA256d tx
merkle (BCH==BTC rule; CTOR only fixes order, already carried) and drops
any header/txs pair that does not cohere -- so a bogus serialized size can
never be folded into the ABLA feed. Mirrors BCHN CheckMerkleRoot at accept.

Extract merkle_hash_pair/compute_merkle_root into coin/merkle.hpp as the
single source shared by the GBT template builder (producer) and the
full-block accept path (consumer); template_builder.hpp now includes it.

Source-only / build-inert (no impl_bch CMake reg). p2pool-merged-v36
surface: NONE -- local parent-chain accept gate, no share/PoW/coinbase
commitment/PPLNS touch.
